Added markdown processing to html
I want to write my blog posts in markdown, but have them render
nicely as html in the site. Investigated multiple markdown processor
options and settled on blackfriday. It seemed fairly well documented
and actively supported.

Doing this made a few changes to the BlogPost struct, and I also had
to update the Dockerfile to now include the go.sum file since we
have an external dependency.